Set the format specifier %lx to long unsigned int
src/via_display.c (xf86DrvMsg): Fix compile warning missmatch by
specifying the format to long unsigned int

+++ src/via_display.c
@@ -1168,11 +1168,11 @@ viaIGA1SetFBStartingAddress(xf86CrtcPtr crtc, int x, int y)
     Base = (y * pScrn->displayWidth + x) * (pScrn->bitsPerPixel / 8);
     DEBUG(xf86DrvMsg(pScrn->scrnIndex, X_INFO,
                         "Base Address: 0x%lx\n",
-                        Base));
+                        (unsigned long)Base));
     Base = (Base + drmmode->front_bo->offset) >> 1;
     xf86DrvMsg(pScrn->scrnIndex, X_INFO,
                 "DRI Base Address: 0x%lx\n",
-                Base);
+                (unsigned long)Base);
 
     hwp->writeCrtc(hwp, 0x0D, Base & 0xFF);
     hwp->writeCrtc(hwp, 0x0C, (Base & 0xFF00) >> 8);
@@ -2321,11 +2321,11 @@ viaIGA2SetFBStartingAddress(xf86CrtcPtr crtc, int x, int y)
     Base = (y * pScrn->displayWidth + x) * (pScrn->bitsPerPixel / 8);
     DEBUG(xf86DrvMsg(pScrn->scrnIndex, X_INFO,
                         "Base Address: 0x%lx\n",
-                        Base));
+                        (unsigned long)Base));
     Base = (Base + drmmode->front_bo->offset) >> 3;
     xf86DrvMsg(pScrn->scrnIndex, X_INFO,
                 "DRI Base Address: 0x%lx\n",
-                Base);
+                (unsigned long)Base);
 
     tmp = hwp->readCrtc(hwp, 0x62) & 0x01;
     tmp |= (Base & 0x7F) << 1;
